fix: CIエラーを修正 - スコア検証ユーティリティ分離とテストアサーション修正
- score-validation/page.tsx の export 関数を lib/score-validation-utils.ts へ移動
  Next.js がページの named export を不正なフィールドとして検出していたため
- tests/admin/score-validation.test.ts: インポートパスを lib/ に更新
- tests/api/resume-review-route.test.ts: ヘッダーアサーションを objectContaining に変更
  実装が X-User-ID / X-User-Token も付加するため
- tests/api/companies-route.test.ts: エラーボディアサーションを実際の動作に合わせて修正
  buildProxyJsonResponse はバックエンドのエラー内容をそのまま返す
